pyproject.toml
[project]
name = "enigmatica"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.1",
]

[dependency-groups]
dev = [
    "ruff>=0.12.1",
]
